SimpleFetcherBolt does not emit to default stream when code 304, added test using Wiremock + http protocol returns empty array when entity is null #279
