+++ core/pom.xml
@@ -27,6 +27,7 @@
 		<httpclient.version>4.4.1</httpclient.version>
 		<snakeyaml.version>1.16</snakeyaml.version>
 		<commons.lang.version>2.6</commons.lang.version>
+		<wiremock.version>1.57</wiremock.version>
 	</properties>
 
 	<build>
@@ -145,7 +146,7 @@
 		<dependency>
 			<groupId>com.github.tomakehurst</groupId>
 			<artifactId>wiremock</artifactId>
-			<version>1.57</version>
+			<version>${wiremock.version}</version>
 			<scope>test</scope>
 		</dependency>
