+++ core/src/main/java/com/digitalpebble/storm/crawler/bolt/SimpleFetcherBolt.java
@@ -346,15 +346,26 @@ public class SimpleFetcherBolt extends BaseRichBolt {
             response.getMetadata().putAll(metadata);
 
             // determine the status based on the status code
-            Status status = Status.fromHTTPCode(response.getStatusCode());
+            final Status status = Status.fromHTTPCode(response.getStatusCode());
+
+            // used when sending to status stream
+            final Values values4status = new Values(urlString,
+                    response.getMetadata(), status);
 
             // if the status is OK emit on default stream
             if (status.equals(Status.FETCHED)) {
-                _collector.emit(
-                        Utils.DEFAULT_STREAM_ID,
-                        input,
-                        new Values(urlString, response.getContent(), response
-                                .getMetadata()));
+                if (response.getStatusCode() == 304) {
+                    // mark this URL as fetched so that it gets
+                    // rescheduled
+                    // but do not try to parse or index
+                    _collector
+                            .emit(com.digitalpebble.storm.crawler.Constants.StatusStreamName,
+                                    input, values4status);
+                } else {
+                    _collector.emit(Utils.DEFAULT_STREAM_ID, input,
+                            new Values(urlString, response.getContent(),
+                                    response.getMetadata()));
+                }
             } else if (status.equals(Status.REDIRECTION)) {
 
                 // find the URL it redirects to
@@ -368,24 +379,19 @@ public class SimpleFetcherBolt extends BaseRichBolt {
                     response.getMetadata().setValue("_redirTo", redirection);
                 }
 
-                // Mark URL as redirected
-                _collector
-                        .emit(com.digitalpebble.storm.crawler.Constants.StatusStreamName,
-                                input,
-                                new Values(urlString, response.getMetadata(),
-                                        status));
-
                 if (allowRedirs && StringUtils.isNotBlank(redirection)) {
                     handleOutlink(input, url, redirection,
                             response.getMetadata());
                 }
+                // Mark URL as redirected
+                _collector
+                        .emit(com.digitalpebble.storm.crawler.Constants.StatusStreamName,
+                                input, values4status);
             } else {
                 // Error
                 _collector
                         .emit(com.digitalpebble.storm.crawler.Constants.StatusStreamName,
-                                input,
-                                new Values(urlString, response.getMetadata(),
-                                        status));
+                                input, values4status);
             }
 
         } catch (Exception exece) {
